+++ pygbif/collection/__init__.py
@@ -1,0 +1,8 @@
+""" 
+GBIF collection APIs methods
+
+* `search`: Search for collections in GRSciColl
+
+"""
+
+from .search import search