+++ pygbif/institution/__init__.py
@@ -1,0 +1,8 @@
+""" 
+GBIF institution APIs methods
+
+* `search`: Search for institutions in GRSciColl
+
+"""
+
+from .search import search