+++ test/test-registry-networks.py
@@ -9,7 +9,7 @@ def test_networks():
     res = registry.networks()
     assert dict == res.__class__
     assert 2 == len(res)
-    assert 14 == len(res["data"])
+    assert len(res["data"]) >= 10 
     assert ["data", "meta"] == sorted(res.keys())
 
 @vcr.use_cassette("test/vcr_cassettes/test_networks_limit.yaml")
@@ -18,7 +18,6 @@ def test_networks_limit():
     res = registry.networks(limit=5)
     assert dict == res.__class__
     assert 5 == len(res["data"])
-
 
 @vcr.use_cassette("test/vcr_cassettes/test_networks_uuid.yaml")
 def test_networks_uuid():